refactor(mesi): split fetchUrl.go into fetch.go + ssrf.go (#120)
